refactor(mobile): extract settings, about and notification screens from their routes
Peel the settings menu, about, notifications, chat-UI and browser-settings
bodies out of their expo-router route files into props-driven components under
mobile/src/settings, leaving each route as a thin mount. No visual change.
